Transfer operator

Entropies.TransferOperator — Type

TransferOperator(r::RectangularBinning) <: BinningProbabilitiesEstimator

A probability estimator based on binning data into rectangular boxes dictated by the binning scheme r, then estimating the transfer (Perron-Frobenius) operator over the bins. Assumes that the input data are sequential.

The transfer operator $P^{N}$is computed as an N-by-N matrix of transition probabilities between the states defined by the partition elements, where N is the number of boxes in the partition that is visited by the orbit/points. The left invariant distribution $\rho^{N}$, where $\rho^{N}P^{N} = \rho^{N}$, approximates the invariant density of the system subject to the partition r. Thus, $\rho^{N}$ can be taken as a probability distribution over the partition elements.

This implementation follows the approach in Diego et al. (2019)[Diego2019].

See also: RectangularBinning.

Utility methods/types

Entropies.TransferOperatorApproximationRectangular — Type

TransferOperatorApproximationRectangular(to, ϵ::RectangularBinning, mini, edgelengths, bins)

The N-by-N matrix to is an approximation to the transfer operator, subject to the partition ϵ, computed over some set of sequentially ordered points.

For convenience, mini and edgelengths provide the minima and box edge lengths along each coordinate axis, as determined by applying ϵ to the points. The coordinates of the (leftmost, if axis is ordered low-high) box corners are given in bins.

Only bins actually visited by the points are considered, and bins give the coordinates of these bins. The element bins[i] correspond to the i-th state of the system, which corresponds to the i-th column/row of the transfer operator to.

See also: RectangularBinning.

Entropies.transferoperator — Function

transferoperator(pts::AbstractDataset{D, T}, ϵ::RectangularBinning) → TransferOperatorApproximationRectangular

Estimate the transfer operator given a set of sequentially ordered points subject to a rectangular partition given by ϵ.

Example

using DynamicalSystems, Entropy
D = 4
ds = Systems.lorenz96(D; F = 32.0)
N, dt = 20000, 0.1
orbit = trajectory(ds, N*dt; dt = dt, Ttr = 10.0)

# Estimate transfer operator over some coarse graining of the orbit.
transferoperator(orbit, RectangularBinning(10))

See also: RectangularBinning.

Entropies.InvariantMeasureEstimate — Type

InvariantMeasureEstimate(to, ρ::Probabilities)

The estimated invariant measure ρ associated with some transfer operator to.

Entropies.invariantmeasure — Function

invariantmeasure(to::TransferOperatorApproximationRectangular) → InvariantMeasureEstimate

Estimate the invariant measure associated with some pre-computed transfer operator.

Example

using DynamicalSystems, Entropy
D = 4
ds = Systems.lorenz96(D; F = 32.0)
N, dt = 20000, 0.1
orbit = trajectory(ds, N*dt; dt = dt, Ttr = 10.0)

# Estimate transfer operator over some coarse graining of the orbit.
to = transferoperator(orbit, RectangularBinning(10))
invariantmeasure(to)

See also: TransferOperatorApproximationRectangular.

Missing docstring.

Missing docstring for binhist(::InvariantMeasureEstimate). Check Documenter's build log for details.

Diego2019Diego, D., Haaga, K. A., & Hannisdal, B. (2019). Transfer entropy computation using the Perron-Frobenius operator. Physical Review E, 99(4), 042212.